Create LogInPageTests,HeaderTabsTests and YourProfilePageTests classes
